Formalize versioning of bundles.
Bundles can now optionally include a codeVersion() field that
returns a float. Someday we can expand this to semver if desired,
but that will be compatible with returning a float.

If putBundle() is called with a bundle whose version is <= the
currently registered bundle, the call is ignored.

As a special case for development, a bundle that includes no
codeVersion() (called "unversioned") is allowed to upgrade another
unversioned bundle.

Fixes #26